WarungTechAI/requirements.txt
# WarungTech AI Assistant - Python Dependencies

# Core AI/LLM
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20

# Web3 & Blockchain (optional)
web3>=6.0.0
eth-utils>=2.0.0

# API & Web
requests>=2.28.0
flask>=2.3.0
flask-cors>=4.0.0

# Environment & Configuration
python-dotenv>=1.0.0

# Data Processing
typing-extensions>=4.5.0

# Development & Testing (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0